+++ src/main/java/com/hakimen/peripherals/MorePeripheralsClient.java
@@ -7,7 +7,9 @@ import com.hakimen.peripherals.registry.ComputerCraftRegister;
 import com.hakimen.peripherals.registry.ContainerRegister;
 import dan200.computercraft.api.client.ComputerCraftAPIClient;
 import dan200.computercraft.api.client.turtle.TurtleUpgradeModeller;
+import dan200.computercraft.shared.turtle.blocks.TurtleBlockEntity;
 import net.minecraft.client.gui.screens.MenuScreens;
+import net.minecraft.world.entity.animal.Turtle;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 
 
@@ -26,6 +28,7 @@ public class MorePeripheralsClient {
             MenuScreens.register(ContainerRegister.keyboardContainer.get(), KeyboardScreen::new);
             MenuScreens.register(ContainerRegister.enderChestInterfaceContainer.get(), EnderChestInterfaceScreen::new);
             MenuScreens.register(ContainerRegister.playerInterfaceContainer.get(), PlayerInterfaceScreen::new);
+            MenuScreens.register(ContainerRegister.scannerContainer.get(), ScannerScreen::new);
         });
 
     }