+++ src/main/java/com/hakimen/peripherals/blocks/ScannerBlock.java
@@ -1,7 +1,9 @@
 package com.hakimen.peripherals.blocks;
 
-import com.hakimen.peripherals.blocks.tile_entities.PlayerInterfaceEntity;
-import com.hakimen.peripherals.client.containers.PlayerInterfaceContainer;
+import com.hakimen.peripherals.blocks.tile_entities.EnderChestInterfaceEntity;
+import com.hakimen.peripherals.blocks.tile_entities.ScannerEntity;
+import com.hakimen.peripherals.client.containers.EnderChestInterfaceContainer;
+import com.hakimen.peripherals.client.containers.ScannerContainer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.network.chat.Component;
@@ -28,11 +30,12 @@ import org.jetbrains.annotations.Nullable;
 
 import javax.annotation.Nonnull;
 
-public class PlayerInterfaceBlock extends Block implements EntityBlock {
+public class ScannerBlock extends Block implements EntityBlock {
 
     public static final DirectionProperty FACING = BlockStateProperties.HORIZONTAL_FACING;
-    public PlayerInterfaceBlock() {
-        super(Properties.copy(Blocks.DIRT).strength(2f,2f).sound(SoundType.STONE));
+    public ScannerBlock() {
+        super(Properties.copy(Blocks.GLASS).strength(2f,2f).sound(SoundType.STONE)
+                .noOcclusion());
         registerDefaultState( getStateDefinition().any()
                 .setValue( FACING, Direction.NORTH ));
     }
@@ -68,23 +71,24 @@ public class PlayerInterfaceBlock extends Block implements EntityBlock {
     @Nullable
     @Override
     public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
-        return new PlayerInterfaceEntity(pos,state);
+        return new ScannerEntity(pos,state);
     }
 
     @Override
     public InteractionResult use(BlockState state, Level level, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hitResult) {
         if (!level.isClientSide) {
             BlockEntity tileEntity = level.getBlockEntity(pos);
-            if (tileEntity instanceof PlayerInterfaceEntity) {
+
+            if (tileEntity instanceof ScannerEntity) {
                 MenuProvider containerProvider = new MenuProvider() {
                     @Override
                     public Component getDisplayName() {
-                        return Component.translatable("gui.peripherals.player_interface.name");
+                        return Component.translatable("gui.peripherals.scanner.name");
                     }
 
                     @Override
                     public AbstractContainerMenu createMenu(int windowId, Inventory playerInventory, Player playerEntity) {
-                        return new PlayerInterfaceContainer(windowId, pos, playerInventory, playerEntity);
+                        return new ScannerContainer(windowId, pos, playerInventory, playerEntity);
                     }
                 };
                 NetworkHooks.openScreen((ServerPlayer) player, containerProvider, tileEntity.getBlockPos());
@@ -95,11 +99,10 @@ public class PlayerInterfaceBlock extends Block implements EntityBlock {
 
         return InteractionResult.SUCCESS;
     }
-
     @Override
     public boolean onDestroyedByPlayer(BlockState state, Level level, BlockPos pos, Player player, boolean willHarvest, FluidState fluid) {
         var blockEntity = level.getBlockEntity(pos);
-        if(blockEntity instanceof PlayerInterfaceEntity entity){
+        if(blockEntity instanceof ScannerEntity entity){
             level.addFreshEntity(new ItemEntity(level,pos.getX(),pos.getY(),pos.getZ(),entity.inventory.getStackInSlot(0)));
         }
         return super.onDestroyedByPlayer(state, level, pos, player, willHarvest, fluid);