+++ src/main/java/com/hakimen/peripherals/blocks/tile_entities/ScannerEntity.java
@@ -1,9 +1,8 @@
 package com.hakimen.peripherals.blocks.tile_entities;
 
-import com.hakimen.peripherals.items.MagneticCardItem;
+import com.hakimen.peripherals.items.PlayerCardItem;
 import com.hakimen.peripherals.registry.BlockEntityRegister;
-import com.hakimen.peripherals.registry.ItemRegister;
-import dan200.computercraft.api.peripheral.IComputerAccess;
+import dan200.computercraft.shared.media.items.PrintoutItem;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
@@ -11,6 +10,9 @@ import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
 import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.WritableBookItem;
+import net.minecraft.world.item.WrittenBookItem;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.capabilities.Capability;
@@ -22,18 +24,18 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.annotation.Nonnull;
-import java.util.ArrayList;
 
 
-public class MagneticCardManiputalorEntity extends BlockEntity {
+public class ScannerEntity extends BlockEntity {
+
 
     public final ItemStackHandler inventory = createHandler();
-    public final ArrayList<IComputerAccess> computers = new ArrayList<>();
+
     private final LazyOptional<IItemHandler> handler = LazyOptional.of(() -> inventory);
 
 
-    public MagneticCardManiputalorEntity(BlockPos pos, BlockState state) {
-        super(BlockEntityRegister.magneticCardManipulator.get(), pos, state);
+    public ScannerEntity(BlockPos pos, BlockState state) {
+        super(BlockEntityRegister.scanner.get(), pos, state);
     }
 
 
@@ -47,6 +49,12 @@ public class MagneticCardManiputalorEntity extends BlockEntity {
     protected void saveAdditional(CompoundTag tag) {
         tag.merge(this.inventory.serializeNBT());
         super.saveAdditional(tag);
+    }
+
+    @Override
+    public void setChanged() {
+        level.sendBlockUpdated(getBlockPos(),getBlockState(),getBlockState(), Block.UPDATE_ALL);
+        super.setChanged();
     }
     @Nullable
     @Override
@@ -62,6 +70,7 @@ public class MagneticCardManiputalorEntity extends BlockEntity {
     public void tick(){
 
     }
+
     @Override
     public @NotNull <T> LazyOptional<T> getCapability(@NotNull Capability<T> cap, Direction dir) {
         if (cap == ForgeCapabilities.ITEM_HANDLER) {
@@ -72,7 +81,6 @@ public class MagneticCardManiputalorEntity extends BlockEntity {
 
 
 
-
     private ItemStackHandler createHandler() {
         return new ItemStackHandler(1) {
             @Override
@@ -80,34 +88,14 @@ public class MagneticCardManiputalorEntity extends BlockEntity {
                 setChanged();
             }
 
-
             @Override
             public boolean isItemValid(int slot, @Nonnull ItemStack stack) {
                 if (slot == 0) {
-                    return stack.getItem() instanceof MagneticCardItem;
+                    return stack.getItem() instanceof PrintoutItem
+                            || stack.getItem() instanceof WrittenBookItem
+                            || stack.getItem() instanceof WritableBookItem;
                 }
                 return false;
-            }
-
-
-            @Override
-            public @NotNull ItemStack extractItem(int slot, int amount, boolean simulate) {
-                if(getStackInSlot(slot).getItem().equals(ItemRegister.magnetic_card.get().asItem()) && !simulate){
-                    for (IComputerAccess c: computers) {
-                        c.queueEvent("card_remove");
-                    }
-                }
-                return super.extractItem(slot, amount, simulate);
-            }
-
-            @Override
-            public void setStackInSlot(int slot, @NotNull ItemStack stack) {
-                if(stack.getItem().equals(ItemRegister.magnetic_card.get().asItem())){
-                    for (IComputerAccess c: computers) {
-                        c.queueEvent("card_insert");
-                    }
-                }
-                super.setStackInSlot(slot, stack);
             }
 
             @Override
@@ -121,6 +109,7 @@ public class MagneticCardManiputalorEntity extends BlockEntity {
                 if(!isItemValid(slot, stack)) {
                     return stack;
                 }
+
                 return super.insertItem(slot, stack, simulate);
             }
         };